refactor(layout): extract HTTP/MCP clients into clients/ layer
services/ 내부에 섞여 있던 외부 I/O 어댑터를 clients/ 로 분리
하여 정책(서비스)과 어댑터(클라이언트) 경계 명확화.

- services/hooks/ltm_client.py -> clients/ltm.py
- services/tts/irodori.py -> clients/irodori.py
- services/channels/desktop_mate_rest.py -> clients/desktop_mate_rest.py
- tests/ 디렉토리도 새 구조에 맞춰 미러링

213 passed 유지.
