+++ src/nanobot_runtime/clients/__init__.py
@@ -1,0 +1,1 @@
+"""External I/O adapters: HTTP/MCP clients."""
