+++ tests/clients/test_desktop_mate_rest.py
